+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/environment-check
 
+## 0.2.0
+
+### Minor Changes
+
+-   08b1c6e: added vscode functionality to environment check
+
+## 0.1.5
+
+### Patch Changes
+
+-   @sap-ux/ui5-config@0.15.1
+
 ## 0.1.4
 
 ### Patch Changes